Application-side error handling
Application creation and callbacks may generate errors (this is
different from repl-errors, which were already being caught and
reported). These errors are stored in the errcatch/errors atom (in
addition to being thrown, errcatch/throw-on-error? flag).
Also, added a feature to reload only the functions, making editing the
code and seeing the changes faster.
However, still it’s easier to externally edit the program for now.