Automatic indentation
Added grammer/newline-parse which determines where it is safe to add
newlines.
Oops the way we do parsing isn't that good: it would have been easier
to more explicitly denote tokens, levels, etc. However,
 this is abstracted away pretty early in the game so we are holding-off
a rewrite of said section.
Bugfix: grammer/_spacers would add newlines in the wrong place (and add
too many) when a child ended in a comment.
Refactor: Got rid of the unused indent file, replaced it with the
indentation.
Bugfix: fixed a dangerous bug where resetting the app would cause files
to be reloaded.
Added the ability to store the repl's text as a string in the var
user/s. Useful for testing indentation, etc on code-as-string.

Mostly self-editing.